quiet unnecessary websocket logs
